Make node sorting fully deterministic
Graphs are built using Go maps which do not provide a determinist
traversal. We sort the nodes to remove the non-determinism, but
the sort fails to provide a fully deterministic ordering in cases
where the attributes being sorted have equal values. Detect those
cases and fallback to a full comparison of all fields.